requirements.txt
python-dotenv~=1.0.0
discord.py~=2.3.2